Use environment variables to control test config
