+++ test/docfx.Test/ModuleInitializer.cs
@@ -11,7 +11,9 @@ namespace Microsoft.Docs.Build
         public static void Initialize()
         {
             Environment.SetEnvironmentVariable("DOCFX_APPDATA_PATH", Path.GetFullPath("appdata"));
-            Environment.SetEnvironmentVariable("DOCFX_GLOBAL_CONFIG_PATH", Path.GetFullPath("docfx.test.yml"));
+            Environment.SetEnvironmentVariable("DOCFX_BASE_URL", "https://docs.com");
+            Environment.SetEnvironmentVariable("DOCFX_OUTPUT__JSON", "true");
+            Environment.SetEnvironmentVariable("DOCFX_OUTPUT__COPY_RESOURCES", "true");
 
             Log.ForceVerbose = true;
             TestUtility.MakeDebugAssertThrowException();
